Refactored board logic to include strategy pattern <organisation
